requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
mysql-connector-python==9.2.0
python-dotenv==1.0.1
PyJWT==2.10.1
bcrypt==4.2.1
Pillow>=10.0.0

